refactor(clawdstrike): extract load and validate from policy.rs
Pure relocation, splitting the giant impl Policy across modules:
- load.rs: new/from_yaml*/to_yaml/resolve_base/from_yaml_with_extends* +
  maybe_verify_loaded_policy + spider_sense_present_fields_from_yaml
- validate/mod.rs: impl Policy::{validate, validate_with_options} +
  glob/domain-glob/placeholder validators
- validate/version.rs: semver parsing + policy_version_supports_* gating
- validate/specs.rs: validate_custom_guards + async-config + provider-spec validators

Sanctioned visibility widening for cross-module access (the inline test module
calls these via the policy module tree):
- POLICY_LOAD_VERIFIER + PolicyLoadVerifier: pub(crate) (read by load::maybe_verify)
- Policy::from_yaml_unvalidated: pub(in crate::policy) (matches prior module-private
  scope; tests call it)

policy_version_supports_origins/broker re-exported at crate::policy path.
mod.rs keeps only impl Policy::create_guards (-> guards.rs next), PolicyGuards,
RuleSet. Now-unused regex/globset/error/placeholders/posture imports dropped from
mod.rs; test-only Error import moved into the test module. validate/mod.rs is 613
lines (the validate_with_options cluster), within the plan's sanctioned range. No
logic edits.
