+++ crates/libs/clawdstrike/src/policy/mod.rs
@@ -1,31 +1,26 @@
 //! Policy configuration and rulesets
 
-use regex::Regex;
 use serde::{Deserialize, Serialize};
-use std::collections::BTreeSet;
-use std::path::Path;
 use std::sync::OnceLock;
 
-use globset::GlobBuilder;
-
-use crate::error::{Error, PolicyFieldError, PolicyValidationError, Result};
+use crate::error::Result;
 use crate::guards::{
     ComputerUseGuard, EgressAllowlistGuard, ForbiddenPathGuard, Guard,
     InputInjectionCapabilityGuard, JailbreakGuard, McpToolGuard, PatchIntegrityGuard,
     PathAllowlistGuard, PromptInjectionGuard, RemoteDesktopSideChannelGuard, SecretLeakGuard,
     ShellCommandGuard,
 };
-use crate::placeholders::env_var_for_placeholder;
-use crate::posture::validate_posture_config;
 
 mod async_config;
 mod broker;
 mod guard_configs;
+mod load;
 mod merge;
 mod origins;
 mod resolver;
 mod settings;
 mod types;
+mod validate;
 
 pub use async_config::{
     AsyncCachePolicyConfig, AsyncCircuitBreakerPolicyConfig, AsyncExecutionMode,
@@ -43,6 +38,7 @@ pub use resolver::{
 };
 pub use settings::{CustomGuardSpec, PolicySettings, VerificationSettings};
 pub use types::{MergeStrategy, Policy, PolicyLoadVerificationInput, PolicyValidationOptions};
+pub use validate::{policy_version_supports_broker, policy_version_supports_origins};
 
 pub(crate) use settings::merge_verification_settings;
 
@@ -54,10 +50,10 @@ pub const POLICY_SCHEMA_VERSION: &str = "1.5.0";
 pub const POLICY_SUPPORTED_SCHEMA_VERSIONS: &[&str] =
     &["1.1.0", "1.2.0", "1.3.0", "1.4.0", "1.5.0"];
 
-type PolicyLoadVerifier =
+pub(crate) type PolicyLoadVerifier =
     dyn Fn(&PolicyLoadVerificationInput) -> Result<()> + Send + Sync + 'static;
 
-static POLICY_LOAD_VERIFIER: OnceLock<Box<PolicyLoadVerifier>> = OnceLock::new();
+pub(crate) static POLICY_LOAD_VERIFIER: OnceLock<Box<PolicyLoadVerifier>> = OnceLock::new();
 
 pub fn install_policy_load_verifier<F>(verifier: F) -> bool
 where
@@ -66,716 +62,7 @@ where
     POLICY_LOAD_VERIFIER.set(Box::new(verifier)).is_ok()
 }
 
-#[cfg(feature = "full")]
-fn spider_sense_present_fields_from_yaml(yaml: &str) -> BTreeSet<String> {
-    fn mapping_get<'a>(map: &'a serde_yaml::Mapping, key: &str) -> Option<&'a serde_yaml::Value> {
-        map.get(serde_yaml::Value::String(key.to_string()))
-    }
-
-    let mut fields = BTreeSet::new();
-    let root = match serde_yaml::from_str::<serde_yaml::Value>(yaml) {
-        Ok(value) => value,
-        Err(_) => return fields,
-    };
-    let Some(root_map) = root.as_mapping() else {
-        return fields;
-    };
-    let Some(guards) = mapping_get(root_map, "guards") else {
-        return fields;
-    };
-    let Some(guards_map) = guards.as_mapping() else {
-        return fields;
-    };
-    let Some(spider_sense) = mapping_get(guards_map, "spider_sense") else {
-        return fields;
-    };
-    let Some(spider_sense_map) = spider_sense.as_mapping() else {
-        return fields;
-    };
-
-    for key in spider_sense_map.keys() {
-        if let Some(name) = key.as_str() {
-            fields.insert(name.to_string());
-        }
-    }
-    fields
-}
-
 impl Policy {
-    /// Create an empty policy
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Load from YAML file
-    pub fn from_yaml_file(path: impl AsRef<Path>) -> Result<Self> {
-        let content = std::fs::read_to_string(path)?;
-        Self::from_yaml(&content)
-    }
-
-    /// Parse from YAML string
-    pub fn from_yaml(yaml: &str) -> Result<Self> {
-        let policy = Self::from_yaml_without_load_verification(yaml)?;
-        if policy.extends.is_none() {
-            maybe_verify_loaded_policy(&policy, Some(&policy), None)?;
-        }
-        Ok(policy)
-    }
-
-    /// Parse from YAML string and validate it without running load-time
-    /// verification hooks.
-    ///
-    /// This is intended for callers that need source metadata before resolving
-    /// `extends`, or that will run load-time verification with richer parent
-    /// context later in the load pipeline.
-    pub fn from_yaml_without_load_verification(yaml: &str) -> Result<Self> {
-        let policy = Self::from_yaml_unvalidated(yaml)?;
-        policy.validate()?;
-        Ok(policy)
-    }
-
-    /// Parse from YAML string, auto-detecting HushSpec or Clawdstrike format.
-    ///
-    /// If the document contains a top-level `hushspec:` key, it is parsed and compiled via
-    /// [`compile_hushspec`](crate::hushspec_compiler::compile_hushspec).
-    /// Otherwise falls through to
-    /// [`from_yaml`](Self::from_yaml).
-    pub fn from_yaml_auto(yaml: &str) -> Result<Self> {
-        if crate::hushspec_compiler::is_hushspec(yaml) {
-            crate::hushspec_compiler::compile_hushspec(yaml)
-        } else {
-            Self::from_yaml(yaml)
-        }
-    }
-
-    fn from_yaml_unvalidated(yaml: &str) -> Result<Self> {
-        let policy: Self = serde_yaml::from_str(yaml)?;
-        #[cfg(feature = "full")]
-        {
-            let mut policy = policy;
-            policy.guards.spider_sense_present_fields = spider_sense_present_fields_from_yaml(yaml);
-            Ok(policy)
-        }
-        #[cfg(not(feature = "full"))]
-        {
-            Ok(policy)
-        }
-    }
-
-    /// Export to YAML string
-    pub fn to_yaml(&self) -> Result<String> {
-        serde_yaml::to_string(self).map_err(Error::from)
-    }
-
-    /// Validate policy semantics and guard configs.
-    ///
-    /// This is a security boundary: invalid regex/glob patterns are treated as errors, not silently ignored.
-    pub fn validate(&self) -> Result<()> {
-        self.validate_with_options(PolicyValidationOptions::default())
-    }
-
-    pub fn validate_with_options(&self, options: PolicyValidationOptions) -> Result<()> {
-        validate_policy_version(&self.version)?;
-
-        let mut errors: Vec<PolicyFieldError> = Vec::new();
-        let require_env = options.require_env;
-        let supports_v1_2_features = policy_version_supports_posture(&self.version);
-
-        if self.posture.is_some() && !supports_v1_2_features {
-            errors.push(PolicyFieldError::new(
-                "posture",
-                "posture requires policy version 1.2.0".to_string(),
-            ));
-        }
-
-        if self.origins.is_some() && !policy_version_supports_origins(&self.version) {
-            errors.push(PolicyFieldError::new(
-                "origins",
-                format!(
-                    "origins block requires schema version >= 1.4.0, got {}",
-                    self.version
-                ),
-            ));
-        }
-
-        if self.broker.is_some() && !policy_version_supports_broker(&self.version) {
-            errors.push(PolicyFieldError::new(
-                "broker",
-                format!(
-                    "broker block requires schema version >= 1.5.0, got {}",
-                    self.version
-                ),
-            ));
-        }
-
-        if let Some(ref broker) = self.broker {
-            if broker.enabled && broker.providers.is_empty() {
-                errors.push(PolicyFieldError::new(
-                    "broker.providers",
-                    "broker.providers must contain at least one provider when broker is enabled"
-                        .to_string(),
-                ));
-            }
-
-            let mut seen_provider_names = std::collections::HashSet::new();
-            for (index, provider) in broker.providers.iter().enumerate() {
-                let prefix = format!("broker.providers[{index}]");
-                if provider.name.trim().is_empty() {
-                    errors.push(PolicyFieldError::new(
-                        format!("{prefix}.name"),
-                        "provider name must be non-empty".to_string(),
-                    ));
-                } else if !seen_provider_names.insert(provider.name.as_str()) {
-                    errors.push(PolicyFieldError::new(
-                        format!("{prefix}.name"),
-                        format!("duplicate broker provider name: {}", provider.name),
-                    ));
-                }
-
-                if provider.host.trim().is_empty() {
-                    errors.push(PolicyFieldError::new(
-                        format!("{prefix}.host"),
-                        "provider host must be non-empty".to_string(),
-                    ));
-                }
-
-                if provider.secret_ref.trim().is_empty() {
-                    errors.push(PolicyFieldError::new(
-                        format!("{prefix}.secret_ref"),
-                        "provider secret_ref must be non-empty".to_string(),
-                    ));
-                }
-
-                if provider.exact_paths.is_empty() {
-                    errors.push(PolicyFieldError::new(
-                        format!("{prefix}.exact_paths"),
-                        "provider exact_paths must contain at least one path".to_string(),
-                    ));
-                }
-
-                if provider.methods.is_empty() {
-                    errors.push(PolicyFieldError::new(
-                        format!("{prefix}.methods"),
-                        "provider methods must contain at least one method".to_string(),
-                    ));
-                }
-
-                for (path_index, path) in provider.exact_paths.iter().enumerate() {
-                    if !path.starts_with('/') {
-                        errors.push(PolicyFieldError::new(
-                            format!("{prefix}.exact_paths[{path_index}]"),
-                            "broker exact path must start with '/'".to_string(),
-                        ));
-                    }
-                }
-            }
-        }
-
-        if let Some(ref origins) = self.origins {
-            let mut seen_ids = std::collections::HashSet::new();
-            for profile in &origins.profiles {
-                if !seen_ids.insert(&profile.id) {
-                    errors.push(PolicyFieldError::new(
-                        "origins.profiles",
-                        format!("duplicate origin profile id: {}", profile.id),
-                    ));
-                }
-                // H3 fix: validate posture references at load time
-                if let Some(ref posture_ref) = profile.posture {
-                    if let Some(ref posture_config) = self.posture {
-                        if !posture_config.states.contains_key(posture_ref) {
-                            errors.push(PolicyFieldError::new(
-                                format!("origins.profiles[{}].posture", profile.id),
-                                format!(
-                                    "references unknown posture state '{}' (available: {:?})",
-                                    posture_ref,
-                                    posture_config.states.keys().collect::<Vec<_>>()
-                                ),
-                            ));
-                        }
-                    } else {
-                        errors.push(PolicyFieldError::new(
-                            format!("origins.profiles[{}].posture", profile.id),
-                            format!(
-                                "references posture state '{}' but no posture config is defined",
-                                posture_ref
-                            ),
-                        ));
-                    }
-                }
-            }
-        }
-
-        if self.guards.path_allowlist.is_some() && !supports_v1_2_features {
-            errors.push(PolicyFieldError::new(
-                "guards.path_allowlist",
-                "path_allowlist requires policy version 1.2.0".to_string(),
-            ));
-        }
-
-        if let Some(posture) = &self.posture {
-            validate_posture_config(posture, &mut errors);
-        }
-
-        if !self.custom_guards.is_empty() {
-            let mut seen: std::collections::HashSet<&str> = std::collections::HashSet::new();
-            for (idx, cg) in self.custom_guards.iter().enumerate() {
-                if cg.id.trim().is_empty() {
-                    errors.push(PolicyFieldError::new(
-                        format!("custom_guards[{}].id", idx),
-                        "id must be non-empty".to_string(),
-                    ));
-                }
-                if !seen.insert(cg.id.as_str()) {
-                    errors.push(PolicyFieldError::new(
-                        format!("custom_guards[{}].id", idx),
-                        format!("duplicate custom guard id: {}", cg.id),
-                    ));
-                }
-                if !cg.config.is_object() {
-                    errors.push(PolicyFieldError::new(
-                        format!("custom_guards[{}].config", idx),
-                        "config must be a JSON object".to_string(),
-                    ));
-                }
-
-                validate_placeholders_in_json(
-                    &mut errors,
-                    &format!("custom_guards[{}].config", idx),
-                    &cg.config,
-                    cg.enabled,
-                    require_env,
-                );
-            }
-        }
-
-        if let Some(cfg) = &self.guards.forbidden_path {
-            if let Some(patterns) = cfg.patterns.as_ref() {
-                validate_globs(&mut errors, "guards.forbidden_path.patterns", patterns);
-                validate_placeholders_in_strings(
-                    &mut errors,
-                    "guards.forbidden_path.patterns",
-                    patterns,
-                    cfg.enabled,
-                    require_env,
-                );
-            }
-            validate_globs(
-                &mut errors,
-                "guards.forbidden_path.exceptions",
-                &cfg.exceptions,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.forbidden_path.exceptions",
-                &cfg.exceptions,
-                cfg.enabled,
-                require_env,
-            );
-            validate_globs(
-                &mut errors,
-                "guards.forbidden_path.additional_patterns",
-                &cfg.additional_patterns,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.forbidden_path.additional_patterns",
-                &cfg.additional_patterns,
-                cfg.enabled,
-                require_env,
-            );
-            validate_globs(
-                &mut errors,
-                "guards.forbidden_path.remove_patterns",
-                &cfg.remove_patterns,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.forbidden_path.remove_patterns",
-                &cfg.remove_patterns,
-                cfg.enabled,
-                require_env,
-            );
-        }
-
-        if let Some(cfg) = &self.guards.path_allowlist {
-            validate_globs(
-                &mut errors,
-                "guards.path_allowlist.file_access_allow",
-                &cfg.file_access_allow,
-            );
-            validate_globs(
-                &mut errors,
-                "guards.path_allowlist.file_write_allow",
-                &cfg.file_write_allow,
-            );
-            validate_globs(
-                &mut errors,
-                "guards.path_allowlist.patch_allow",
-                &cfg.patch_allow,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.path_allowlist.file_access_allow",
-                &cfg.file_access_allow,
-                cfg.enabled,
-                require_env,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.path_allowlist.file_write_allow",
-                &cfg.file_write_allow,
-                cfg.enabled,
-                require_env,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.path_allowlist.patch_allow",
-                &cfg.patch_allow,
-                cfg.enabled,
-                require_env,
-            );
-        }
-
-        if let Some(cfg) = &self.guards.egress_allowlist {
-            validate_domain_globs(&mut errors, "guards.egress_allowlist.allow", &cfg.allow);
-            validate_domain_globs(&mut errors, "guards.egress_allowlist.block", &cfg.block);
-            validate_domain_globs(
-                &mut errors,
-                "guards.egress_allowlist.additional_allow",
-                &cfg.additional_allow,
-            );
-            validate_domain_globs(
-                &mut errors,
-                "guards.egress_allowlist.additional_block",
-                &cfg.additional_block,
-            );
-            validate_domain_globs(
-                &mut errors,
-                "guards.egress_allowlist.remove_allow",
-                &cfg.remove_allow,
-            );
-            validate_domain_globs(
-                &mut errors,
-                "guards.egress_allowlist.remove_block",
-                &cfg.remove_block,
-            );
-
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.egress_allowlist.allow",
-                &cfg.allow,
-                cfg.enabled,
-                require_env,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.egress_allowlist.block",
-                &cfg.block,
-                cfg.enabled,
-                require_env,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.egress_allowlist.additional_allow",
-                &cfg.additional_allow,
-                cfg.enabled,
-                require_env,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.egress_allowlist.additional_block",
-                &cfg.additional_block,
-                cfg.enabled,
-                require_env,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.egress_allowlist.remove_allow",
-                &cfg.remove_allow,
-                cfg.enabled,
-                require_env,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.egress_allowlist.remove_block",
-                &cfg.remove_block,
-                cfg.enabled,
-                require_env,
-            );
-        }
-
-        if let Some(cfg) = &self.guards.secret_leak {
-            for (idx, p) in cfg.patterns.iter().enumerate() {
-                validate_placeholders_in_string(
-                    &mut errors,
-                    &format!("guards.secret_leak.patterns[{}].name", idx),
-                    &p.name,
-                    cfg.enabled,
-                    require_env,
-                );
-                validate_placeholders_in_string(
-                    &mut errors,
-                    &format!("guards.secret_leak.patterns[{}].pattern", idx),
-                    &p.pattern,
-                    cfg.enabled,
-                    require_env,
-                );
-
-                if let Err(e) = Regex::new(&p.pattern) {
-                    errors.push(PolicyFieldError::new(
-                        format!("guards.secret_leak.patterns[{}].pattern", idx),
-                        format!("invalid regex ({}): {}", p.name, e),
-                    ));
-                }
-            }
-            for (idx, p) in cfg.additional_patterns.iter().enumerate() {
-                validate_placeholders_in_string(
-                    &mut errors,
-                    &format!("guards.secret_leak.additional_patterns[{}].name", idx),
-                    &p.name,
-                    cfg.enabled,
-                    require_env,
-                );
-                validate_placeholders_in_string(
-                    &mut errors,
-                    &format!("guards.secret_leak.additional_patterns[{}].pattern", idx),
-                    &p.pattern,
-                    cfg.enabled,
-                    require_env,
-                );
-
-                if let Err(e) = Regex::new(&p.pattern) {
-                    errors.push(PolicyFieldError::new(
-                        format!("guards.secret_leak.additional_patterns[{}].pattern", idx),
-                        format!("invalid regex ({}): {}", p.name, e),
-                    ));
-                }
-            }
-            validate_globs(
-                &mut errors,
-                "guards.secret_leak.skip_paths",
-                &cfg.skip_paths,
-            );
-            validate_placeholders_in_strings(
-                &mut errors,
-                "guards.secret_leak.skip_paths",
-                &cfg.skip_paths,
-                cfg.enabled,
-                require_env,
-            );
-        }
-
-        if let Some(cfg) = &self.guards.patch_integrity {
-            for (idx, pattern) in cfg.forbidden_patterns.iter().enumerate() {
-                if let Err(e) = Regex::new(pattern) {
-                    errors.push(PolicyFieldError::new(
-                        format!("guards.patch_integrity.forbidden_patterns[{}]", idx),
-                        format!("invalid regex: {}", e),
-                    ));
-                }
-                validate_placeholders_in_string(
-                    &mut errors,
-                    &format!("guards.patch_integrity.forbidden_patterns[{}]", idx),
-                    pattern,
-                    cfg.enabled,
-                    require_env,
-                );
-            }
-        }
-
-        if let Some(cfg) = &self.guards.shell_command {
-            for (idx, pattern) in cfg.forbidden_patterns.iter().enumerate() {
-                if let Err(e) = Regex::new(pattern) {
-                    errors.push(PolicyFieldError::new(
-                        format!("guards.shell_command.forbidden_patterns[{}]", idx),
-                        format!("invalid regex: {}", e),
-                    ));
-                }
-                validate_placeholders_in_string(
-                    &mut errors,
-                    &format!("guards.shell_command.forbidden_patterns[{}]", idx),
-                    pattern,
-                    cfg.enabled,
-                    require_env,
-                );
-            }
-        }
-
-        if let Some(cfg) = &self.guards.prompt_injection {
-            if cfg.max_scan_bytes == 0 {
-                errors.push(PolicyFieldError::new(
-                    "guards.prompt_injection.max_scan_bytes".to_string(),
-                    "max_scan_bytes must be > 0".to_string(),
-                ));
-            }
-            if !cfg.block_at_or_above.at_least(cfg.warn_at_or_above) {
-                errors.push(PolicyFieldError::new(
-                    "guards.prompt_injection.warn_at_or_above".to_string(),
-                    "warn_at_or_above must be <= block_at_or_above".to_string(),
-                ));
-            }
-        }
-
-        if !self.guards.custom.is_empty() {
-            validate_custom_guards(&mut errors, &self.guards.custom, require_env);
-        }
-
-        if errors.is_empty() {
-            Ok(())
-        } else {
-            Err(PolicyValidationError::new(errors).into())
-        }
-    }
-
-    /// Resolve a base policy by name or path
-    ///
-    /// Tries built-in ruleset names first,
-    /// then falls back to loading from file path.
-    pub fn resolve_base(name_or_path: &str) -> Result<Self> {
-        // Try built-in rulesets first
-        if let Some(ruleset) = RuleSet::by_name(name_or_path)? {
-            return Ok(ruleset.policy);
-        }
-
-        // Try loading from file
-        let path = std::path::Path::new(name_or_path);
-        if path.exists() {
-            return Self::from_yaml_file(path);
-        }
-
-        Err(Error::ConfigError(format!(
-            "Unknown ruleset or file not found: {}",
-            name_or_path
-        )))
-    }
-
-    /// Load from YAML string with extends resolution
-    ///
-    /// If the policy has an `extends` field, loads the base and merges.
-    /// Detects circular dependencies.
-    pub fn from_yaml_with_extends(yaml: &str, base_path: Option<&Path>) -> Result<Self> {
-        let resolver = LocalPolicyResolver::new();
-        Self::from_yaml_with_extends_resolver(yaml, base_path, &resolver)
-    }
-
-    /// Load from YAML string with extends resolution using a custom resolver.
-    ///
-    /// This allows callers to support remote `extends` while keeping the default path
-    /// filesystem-only.
-    pub fn from_yaml_with_extends_resolver(
-        yaml: &str,
-        base_path: Option<&Path>,
-        resolver: &impl PolicyResolver,
-    ) -> Result<Self> {
-        let location = base_path
-            .map(|p| PolicyLocation::File(p.to_path_buf()))
-            .unwrap_or(PolicyLocation::None);
-
-        Self::from_yaml_with_extends_location_resolver(yaml, location, resolver)
-    }
-
-    /// Like `from_yaml_with_extends` but with an explicit source location.
-    pub fn from_yaml_with_extends_location_resolver(
-        yaml: &str,
-        location: PolicyLocation,
-        resolver: &impl PolicyResolver,
-    ) -> Result<Self> {
-        Self::from_yaml_with_extends_internal_resolver(
-            yaml,
-            location,
-            resolver,
-            &mut std::collections::HashSet::new(),
-            0,
-            PolicyValidationOptions::default(),
-        )
-    }
-
-    fn from_yaml_with_extends_internal_resolver(
-        yaml: &str,
-        location: PolicyLocation,
-        resolver: &impl PolicyResolver,
-        visited: &mut std::collections::HashSet<String>,
-        depth: usize,
-        validation: PolicyValidationOptions,
-    ) -> Result<Self> {
-        if depth > crate::core::cycle::MAX_POLICY_EXTENDS_DEPTH {
-            return Err(Error::ConfigError(format!(
-                "Policy extends depth exceeded (limit: {})",
-                crate::core::cycle::MAX_POLICY_EXTENDS_DEPTH
-            )));
-        }
-
-        let child = Policy::from_yaml_unvalidated(yaml)?;
-
-        if let Some(ref extends) = child.extends {
-            let resolved = resolver.resolve(extends, &location)?;
-
-            match crate::core::cycle::check_extends_cycle(&resolved.key, visited, depth + 1) {
-                crate::core::cycle::CycleCheckResult::Ok => {}
-                crate::core::cycle::CycleCheckResult::DepthExceeded { limit, .. } => {
-                    return Err(Error::ConfigError(format!(
-                        "Policy extends depth exceeded (limit: {})",
-                        limit
-                    )));
-                }
-                crate::core::cycle::CycleCheckResult::CycleDetected { .. } => {
-                    return Err(Error::ConfigError(format!(
-                        "Circular policy extension detected: {}",
-                        extends
-                    )));
-                }
-            }
-            visited.insert(resolved.key);
-
-            let base = Self::from_yaml_with_extends_internal_resolver(
-                &resolved.yaml,
-                resolved.location,
-                resolver,
-                visited,
-                depth + 1,
-                validation,
-            )?;
-
-            let merged = base.merge(&child);
-            merged.validate_with_options(validation)?;
-            maybe_verify_loaded_policy(&merged, Some(&child), Some(&base))?;
-            Ok(merged)
-        } else {
-            child.validate_with_options(validation)?;
-            maybe_verify_loaded_policy(&child, Some(&child), None)?;
-            Ok(child)
-        }
-    }
-
-    /// Load from YAML string with extends resolution using a custom resolver and validation options.
-    pub fn from_yaml_with_extends_resolver_with_validation_options(
-        yaml: &str,
-        base_path: Option<&Path>,
-        resolver: &impl PolicyResolver,
-        validation: PolicyValidationOptions,
-    ) -> Result<Self> {
-        let location = base_path
-            .map(|p| PolicyLocation::File(p.to_path_buf()))
-            .unwrap_or(PolicyLocation::None);
-
-        Self::from_yaml_with_extends_internal_resolver(
-            yaml,
-            location,
-            resolver,
-            &mut std::collections::HashSet::new(),
-            0,
-            validation,
-        )
-    }
-
-    /// Load from YAML file with extends resolution
-    pub fn from_yaml_file_with_extends(path: impl AsRef<Path>) -> Result<Self> {
-        let path = path.as_ref();
-        let content = std::fs::read_to_string(path)?;
-        Self::from_yaml_with_extends(&content, Some(path))
-    }
-
     /// Create guards from this policy
     pub(crate) fn create_guards(&self) -> PolicyGuards {
         PolicyGuards {
@@ -853,469 +140,6 @@ impl Policy {
                 .unwrap_or_default(),
         }
     }
-}
-
-fn maybe_verify_loaded_policy(
-    effective_policy: &Policy,
-    source_policy: Option<&Policy>,
-    parent_policy: Option<&Policy>,
-) -> Result<()> {
-    let settings = effective_policy.settings.effective_verification();
-    if !settings.enabled {
-        return Ok(());
-    }
-
-    let Some(verifier) = POLICY_LOAD_VERIFIER.get() else {
-        let message = "policy.settings.verification is enabled, but no load-time policy verifier is registered";
-        if settings.strict {
-            return Err(Error::ConfigError(message.to_string()));
-        }
-        tracing::warn!("{}", message);
-        return Ok(());
-    };
-
-    verifier(&PolicyLoadVerificationInput {
-        effective_policy: effective_policy.clone(),
-        source_policy: source_policy.cloned(),
-        parent_policy: parent_policy.cloned(),
-    })
-}
-
-fn validate_policy_version(version: &str) -> Result<()> {
-    if parse_semver_strict(version).is_none() {
-        return Err(Error::InvalidPolicyVersion {
-            version: version.to_string(),
-        });
-    }
-
-    if !POLICY_SUPPORTED_SCHEMA_VERSIONS.contains(&version) {
-        return Err(Error::UnsupportedPolicyVersion {
-            found: version.to_string(),
-            supported: POLICY_SUPPORTED_SCHEMA_VERSIONS.join(", "),
-        });
-    }
-
-    Ok(())
-}
-
-fn parse_semver_strict(version: &str) -> Option<(u64, u64, u64)> {
-    let mut parts = version.split('.');
-    let major = parse_semver_part(parts.next()?)?;
-    let minor = parse_semver_part(parts.next()?)?;
-    let patch = parse_semver_part(parts.next()?)?;
-    if parts.next().is_some() {
-        return None;
-    }
-
-    Some((major, minor, patch))
-}
-
-fn semver_at_least(version: &str, minimum: (u64, u64, u64)) -> bool {
-    let Some(found) = parse_semver_strict(version) else {
-        return false;
-    };
-    found >= minimum
-}
-
-fn policy_version_supports_posture(version: &str) -> bool {
-    semver_at_least(version, (1, 2, 0))
-}
-
-/// Returns true if the given schema version supports origin-aware enforcement.
-pub fn policy_version_supports_origins(version: &str) -> bool {
-    semver_at_least(version, (1, 4, 0))
-}
-
-/// Returns true if the given schema version supports brokered egress policy.
-pub fn policy_version_supports_broker(version: &str) -> bool {
-    semver_at_least(version, (1, 5, 0))
-}
-
-fn parse_semver_part(part: &str) -> Option<u64> {
-    if part.is_empty() {
-        return None;
-    }
-    if part.len() > 1 && part.starts_with('0') {
-        return None;
-    }
-    if !part.bytes().all(|b| b.is_ascii_digit()) {
-        return None;
-    }
-    part.parse().ok()
-}
-
-fn validate_globs(errors: &mut Vec<PolicyFieldError>, field: &str, patterns: &[String]) {
-    for (idx, pattern) in patterns.iter().enumerate() {
-        if let Err(e) = glob::Pattern::new(pattern) {
-            errors.push(PolicyFieldError::new(
-                format!("{}[{}]", field, idx),
-                format!("invalid glob {:?}: {}", pattern, e),
-            ));
-        }
-    }
-}
-
-fn validate_domain_globs(errors: &mut Vec<PolicyFieldError>, field: &str, patterns: &[String]) {
-    for (idx, pattern) in patterns.iter().enumerate() {
-        if let Err(e) = GlobBuilder::new(pattern)
-            .case_insensitive(true)
-            .literal_separator(true)
-            .build()
-        {
-            errors.push(PolicyFieldError::new(
-                format!("{}[{}]", field, idx),
-                format!("invalid domain glob {:?}: {}", pattern, e),
-            ));
-        }
-    }
-}
-
-fn validate_placeholders_in_string(
-    errors: &mut Vec<PolicyFieldError>,
-    field: &str,
-    value: &str,
-    enabled: bool,
-    require_env: bool,
-) {
-    if !enabled {
-        return;
-    }
-
-    let mut i = 0usize;
-    while let Some(start_rel) = value[i..].find("${") {
-        let start = i + start_rel;
-        let after = start + 2;
-
-        let Some(end_rel) = value[after..].find('}') else {
-            break;
-        };
-        let end = after + end_rel;
-
-        let raw = &value[after..end];
-        let env_name = match env_var_for_placeholder(raw) {
-            Ok(v) => v,
-            Err(msg) => {
-                errors.push(PolicyFieldError::new(field, msg));
-                i = end + 1;
-                continue;
-            }
-        };
-
-        if require_env && std::env::var(&env_name).is_err() {
-            errors.push(PolicyFieldError::new(
-                field,
-                format!("missing environment variable {}", env_name),
-            ));
-        }
-
-        i = end + 1;
-    }
-}
-
-fn validate_placeholders_in_strings(
-    errors: &mut Vec<PolicyFieldError>,
-    field: &str,
-    values: &[String],
-    enabled: bool,
-    require_env: bool,
-) {
-    if !enabled {
-        return;
-    }
-
-    for (idx, v) in values.iter().enumerate() {
-        validate_placeholders_in_string(
-            errors,
-            &format!("{}[{}]", field, idx),
-            v,
-            enabled,
-            require_env,
-        );
-    }
-}
-
-fn validate_placeholders_in_json(
-    errors: &mut Vec<PolicyFieldError>,
-    field: &str,
-    value: &serde_json::Value,
-    enabled: bool,
-    require_env: bool,
-) {
-    if !enabled {
-        return;
-    }
-
-    match value {
-        serde_json::Value::String(s) => {
-            validate_placeholders_in_string(errors, field, s, enabled, require_env);
-        }
-        serde_json::Value::Array(items) => {
-            for (idx, v) in items.iter().enumerate() {
-                validate_placeholders_in_json(
-                    errors,
-                    &format!("{}[{}]", field, idx),
-                    v,
-                    enabled,
-                    require_env,
-                );
-            }
-        }
-        serde_json::Value::Object(map) => {
-            for (k, v) in map {
-                validate_placeholders_in_json(
-                    errors,
-                    &format!("{}.{}", field, k),
-                    v,
-                    enabled,
-                    require_env,
-                );
-            }
-        }
-        _ => {}
-    }
-}
-
-fn validate_custom_guards(
-    errors: &mut Vec<PolicyFieldError>,
-    guards: &[CustomGuardSpec],
-    require_env: bool,
-) {
-    for (idx, spec) in guards.iter().enumerate() {
-        let base = format!("guards.custom[{}]", idx);
-
-        validate_placeholders_in_string(
-            errors,
-            &format!("{base}.package"),
-            &spec.package,
-            spec.enabled,
-            require_env,
-        );
-        if let Some(v) = spec.version.as_ref() {
-            validate_placeholders_in_string(
-                errors,
-                &format!("{base}.version"),
-                v,
-                spec.enabled,
-                require_env,
-            );
-        }
-        if let Some(r) = spec.registry.as_ref() {
-            validate_placeholders_in_string(
-                errors,
-                &format!("{base}.registry"),
-                r,
-                spec.enabled,
-                require_env,
-            );
-        }
-
-        validate_placeholders_in_json(
-            errors,
-            &format!("{base}.config"),
-            &spec.config,
-            spec.enabled,
-            require_env,
-        );
-
-        if let Some(async_cfg) = spec.async_config.as_ref() {
-            validate_async_policy_config(errors, &format!("{base}.async"), async_cfg);
-        }
-
-        if !spec.enabled {
-            continue;
-        }
-
-        match spec.package.as_str() {
-            "clawdstrike-virustotal" => validate_virustotal_spec(errors, &base, &spec.config),
-            "clawdstrike-safe-browsing" => validate_safe_browsing_spec(errors, &base, &spec.config),
-            "clawdstrike-snyk" => validate_snyk_spec(errors, &base, &spec.config),
-            "clawdstrike-spider-sense" => {
-                // Spider-Sense config is validated at guard construction time (fail-closed).
-                // Deprecated: use guards.spider_sense instead of guards.custom.
-            }
-            other => errors.push(PolicyFieldError::new(
-                format!("{base}.package"),
-                format!("unsupported custom guard package: {}", other),
-            )),
-        }
-    }
-}
-
-fn validate_async_policy_config(
-    errors: &mut Vec<PolicyFieldError>,
-    base: &str,
-    cfg: &AsyncGuardPolicyConfig,
-) {
-    if let Some(timeout_ms) = cfg.timeout_ms {
-        if !(100..=300_000).contains(&timeout_ms) {
-            errors.push(PolicyFieldError::new(
-                format!("{}.timeout_ms", base),
-                "timeout_ms must be between 100 and 300000".to_string(),
-            ));
-        }
-    }
-
-    if let Some(cache) = cfg.cache.as_ref() {
-        if let Some(ttl) = cache.ttl_seconds {
-            if ttl == 0 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.cache.ttl_seconds", base),
-                    "ttl_seconds must be >= 1".to_string(),
-                ));
-            }
-        }
-        if let Some(max) = cache.max_size_mb {
-            if max == 0 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.cache.max_size_mb", base),
-                    "max_size_mb must be >= 1".to_string(),
-                ));
-            }
-        }
-    }
-
-    if let Some(rl) = cfg.rate_limit.as_ref() {
-        if let Some(rps) = rl.requests_per_second {
-            if rps <= 0.0 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.rate_limit.requests_per_second", base),
-                    "requests_per_second must be > 0".to_string(),
-                ));
-            }
-        }
-        if let Some(rpm) = rl.requests_per_minute {
-            if rpm <= 0.0 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.rate_limit.requests_per_minute", base),
-                    "requests_per_minute must be > 0".to_string(),
-                ));
-            }
-        }
-        if rl.requests_per_second.is_some() && rl.requests_per_minute.is_some() {
-            errors.push(PolicyFieldError::new(
-                format!("{}.rate_limit", base),
-                "specify only one of requests_per_second or requests_per_minute".to_string(),
-            ));
-        }
-        if let Some(burst) = rl.burst {
-            if burst == 0 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.rate_limit.burst", base),
-                    "burst must be >= 1".to_string(),
-                ));
-            }
-        }
-    }
-
-    if let Some(cb) = cfg.circuit_breaker.as_ref() {
-        if let Some(thr) = cb.failure_threshold {
-            if thr == 0 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.circuit_breaker.failure_threshold", base),
-                    "failure_threshold must be >= 1".to_string(),
-                ));
-            }
-        }
-        if let Some(ms) = cb.reset_timeout_ms {
-            if ms < 1000 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.circuit_breaker.reset_timeout_ms", base),
-                    "reset_timeout_ms must be >= 1000".to_string(),
-                ));
-            }
-        }
-        if let Some(thr) = cb.success_threshold {
-            if thr == 0 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.circuit_breaker.success_threshold", base),
-                    "success_threshold must be >= 1".to_string(),
-                ));
-            }
-        }
-    }
-
-    if let Some(retry) = cfg.retry.as_ref() {
-        if let Some(mult) = retry.multiplier {
-            if mult < 1.0 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.retry.multiplier", base),
-                    "multiplier must be >= 1".to_string(),
-                ));
-            }
-        }
-        if let Some(ms) = retry.initial_backoff_ms {
-            if ms < 100 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.retry.initial_backoff_ms", base),
-                    "initial_backoff_ms must be >= 100".to_string(),
-                ));
-            }
-        }
-        if let Some(ms) = retry.max_backoff_ms {
-            if ms < 100 {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.retry.max_backoff_ms", base),
-                    "max_backoff_ms must be >= 100".to_string(),
-                ));
-            }
-        }
-        if let (Some(init), Some(max)) = (retry.initial_backoff_ms, retry.max_backoff_ms) {
-            if max < init {
-                errors.push(PolicyFieldError::new(
-                    format!("{}.retry.max_backoff_ms", base),
-                    "max_backoff_ms must be >= initial_backoff_ms".to_string(),
-                ));
-            }
-        }
-    }
-}
-
-fn require_config_string(
-    errors: &mut Vec<PolicyFieldError>,
-    base: &str,
-    config: &serde_json::Value,
-    key: &str,
-) -> Option<String> {
-    let serde_json::Value::Object(map) = config else {
-        errors.push(PolicyFieldError::new(
-            format!("{base}.config"),
-            "config must be an object".to_string(),
-        ));
-        return None;
-    };
-
-    match map.get(key) {
-        Some(serde_json::Value::String(s)) if !s.trim().is_empty() => Some(s.clone()),
-        _ => {
-            errors.push(PolicyFieldError::new(
-                format!("{base}.config.{key}"),
-                "missing/invalid required string".to_string(),
-            ));
-            None
-        }
-    }
-}
-
-fn validate_virustotal_spec(
-    errors: &mut Vec<PolicyFieldError>,
-    base: &str,
-    config: &serde_json::Value,
-) {
-    let _ = require_config_string(errors, base, config, "api_key");
-}
-
-fn validate_safe_browsing_spec(
-    errors: &mut Vec<PolicyFieldError>,
-    base: &str,
-    config: &serde_json::Value,
-) {
-    let _ = require_config_string(errors, base, config, "api_key");
-    let _ = require_config_string(errors, base, config, "client_id");
-}
-
-fn validate_snyk_spec(errors: &mut Vec<PolicyFieldError>, base: &str, config: &serde_json::Value) {
-    let _ = require_config_string(errors, base, config, "api_token");
-    let _ = require_config_string(errors, base, config, "org_id");
 }
 
 /// Guards instantiated from a policy
@@ -1434,6 +258,7 @@ mod tests {
     #![allow(clippy::expect_used, clippy::unwrap_used)]
 
     use super::*;
+    use crate::error::Error;
     use crate::guards::{ForbiddenPathConfig, SecretLeakConfig};
     use crate::origin::{OriginProvider, ProvenanceConfidence, SpaceType, Visibility};
     use std::sync::Mutex;
